refactor(cli): extract search results renderer to _search_render
Moves the 180-LOC _render_search_rows function (which shapes
ranked-list answers, the schema-explorer fallback, and inventory rows
into Rich tables with optional debug columns) out of
amx/cli_support/commands/search.py into a focused
amx/cli_support/_search_render.py.

search.py re-exports the function so existing test imports keep
working. The four CLI-integration tests that monkeypatched
amx.cli_support.commands.search.console now target the renderer's
own module (amx.cli_support._search_render.console) since that's
where the print calls happen.

search.py: 1466 -> 1283 LOC (-183, -12%).
